python/requirements.txt
# Python Dependencies
# Required for animation/script.py
Pillow>=10.0.0
